requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15.1
drf-spectacular>=0.27.0
PyYAML>=6.0
